requirements/base.txt
Django>=5.1,<6.1
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.0,<4.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.3,<6.0
firebase-admin>=6.5,<7.0
Pillow
django-cleanup
celery
redis